# docs/openapi-bundled.yml
openapi: 3.1.0
info:
  title: Task API
  version: 1.0.0
servers:
  - url: http://localhost:3000
tags:
  - name: Tasks
    description: tasks-related endpoints
paths:
  /tasks:
    get:
      summary: List all tasks
      tags:
        - Tasks
      parameters:
        - name: completed
          in: query
          required: false
          description: Filter tasks by completion status
          schema:
            type: boolean
            example: true
        - name: limit
          in: query
          required: false
          description: Maximum number of tasks to return
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          readOnly: true
          type: integer
          example: 1
        title:
          type: string
          example: Buy groceries
        completed:
          type: boolean
          example: false
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: string
            example: Title is required